+++ src/apps/write_benchmark_graph.cpp
@@ -50,7 +50,7 @@ transport::Graph make_grid_graph(uint32_t width, uint32_t height) {
     graph.coords = std::move(coords);
     graph.offsets.assign(static_cast<size_t>(vertices) + 1, 0);
     for (uint32_t v = 0; v < vertices; ++v) {
-        graph.offsets[v + 1] = graph.offsets[v] + static_cast<uint64_t>(rows[v].size());
+        graph.offsets[v + 1] = graph.offsets[v] + rows[v].size();
     }
     for (const std::vector<transport::Edge> &row : rows) {
         graph.edges.insert(graph.edges.end(), row.begin(), row.end());
